// frontend/src/app/api/model/task.ts
/**
 * Engager Recruitment API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Task { 
    id: string;
    createdAt: string;
    title: string;
    due?: string;
    completedAt?: string;
    status: Task.StatusEnum;
}
export namespace Task {
    export const StatusEnum = {
        Open: 'OPEN',
        Done: 'DONE'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}